data-ingestion-gateway/requirements.txt
Flask==2.2.3
werkzeug==2.2.3
kafka-python==2.0.2
